# fix(perf): Compare char codes for 100% speed up