--- a/js/.eslintrc.cjs
+++ b/js/.eslintrc.cjs
@@ -9,7 +9,7 @@ module.exports = {
     },
     "extends": "eslint:recommended",
     "parserOptions": {
-        "ecmaVersion": 2020,
+        "ecmaVersion": "latest",
         "sourceType": "module"
     },
     "rules": {