--- a/js/package.json
+++ b/js/package.json
@@ -59,6 +59,6 @@
     "rollup-plugin-copy": "^3.4.0",
     "rollup-plugin-dts": "^6.1.1",
     "sinon": "^15.0.1",
-    "typescript": "^4.9.5"
+    "typescript": "^5.5.4"
   }
 }