--- a/js/tsconfig.json
+++ b/js/tsconfig.json
@@ -5,6 +5,7 @@
     "compilerOptions": {
         "declaration": true,
         "emitDeclarationOnly": true,
+        "allowImportingTsExtensions": true,
         "checkJs": true,
         "target": "ES2015",
         "outDir": "temp",