Updated the tokenizer to compare character codes instead of strings. This more than doubles the speed of the parser.

Before:

```
  Momoa JS     x 18.69 ops/sec ±12.89% (37 runs sampled)
  Momoa WASM   x 21.50 ops/sec ±9.04% (40 runs sampled)
  json-to-ast  x 24.13 ops/sec ±9.96% (46 runs sampled)
  parseJson.js x 64.53 ops/sec ±7.02% (58 runs sampled)
```

After

```
  Momoa JS     x 37.91 ops/sec ±11.62% (43 runs sampled)
  Momoa WASM   x 30.57 ops/sec ±8.36% (43 runs sampled)
  json-to-ast  x 24.75 ops/sec ±9.08% (44 runs sampled)
  parseJson.js x 71.66 ops/sec ±3.27% (62 runs sampled)
```